+++ src/main/java/tbs/newgenteacherselect/service/AdminService.java
@@ -3,5 +3,6 @@ package tbs.newgenteacherselect.service;
 import tbs.pojo.BasicUser;
 
 public interface AdminService {
-    void saveAdmin(String token, String name, String password, String phone, int department);
+    void saveAdmin(String token, String name, String password, String phone, long
+            department);
 }
